+++ src/captchakit/challenges/__init__.py
@@ -2,9 +2,9 @@
 
 from __future__ import annotations
 
-from pycaptcha.challenges.base import Challenge, ChallengeFactory, ChallengeSpec
-from pycaptcha.challenges.math import MathChallengeFactory
-from pycaptcha.challenges.text import TextChallengeFactory
+from captchakit.challenges.base import Challenge, ChallengeFactory, ChallengeSpec
+from captchakit.challenges.math import MathChallengeFactory
+from captchakit.challenges.text import TextChallengeFactory
 
 __all__ = [
     "Challenge",