+++ src/captchakit/renderers/__init__.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from pycaptcha.renderers.base import Renderer
-from pycaptcha.renderers.image import ImageRenderer
+from captchakit.renderers.base import Renderer
+from captchakit.renderers.image import ImageRenderer
 
 __all__ = ["ImageRenderer", "Renderer"]