+++ src/captchakit/storage/__init__.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from pycaptcha.storage.base import Storage
-from pycaptcha.storage.memory import MemoryStorage
+from captchakit.storage.base import Storage
+from captchakit.storage.memory import MemoryStorage
 
 __all__ = ["MemoryStorage", "Storage"]